Implement optional hostname canonicalization for kerberos (SPNEGO)
This commit implements an additional property,
useCanonicalHostname, which is supported by the AbstractSpnegoAuthSupplier,
and therefore available to SpnegoAuthSupplier and
KerberosAuthOutInterceptor.

The default behavior is not modified.  If useCanonicalHostname is set,
then the service principal will be created using the canonical name of the
remote host.

For example, if there is a CNAME DNS record for host.example.com which
refers to an A record host-a-b-c.example.com, the default behavior would
be to generate a service principal HTTP/host.example.com. When
useCanonicalHostname is set to true, the service principal would be HTTP/
host-a-b-c.example.com. Most browsers and many other HTTP user agents
perform this kind of canonicalization.


Former-commit-id: dfa59762b50f26263716bae769bc4078d4059c37