[CXF-7527] getMatchedURIs to avoid duplicate URIs
The UriInfo.getMatchedURIs method currently returns "duplicate" URI
entries when called from a method that is not annotated with an
@Path annotation. The "duplicate" URI is basically the same URI but
with a trailing slash. This was first discovered when using sub
resource locators, but could also occur in simpler scenarios (like
a resource method annotated with @GET but not @Path).

When comparing against the javadoc and the RI, it looks like the
URIs we were returning started with a slash, where the RI and
javadoc examples do not. This commit removes the starting slash as
well.


Former-commit-id: 9e0f7ea38b6afe8f4180082d3a82d9f969a05df3