Handle ResponseExceptionMappers
We should not need to call handleMapper for the ResponseExceptionMappers
 - they already require the parameterized type to be a subclass of
Throwable. And the handleMapper method checks whether the Throwable
class can be assigned to the type in the user's REM - which it never
will be (unless they define ResponseExceptionMapper<Throwable>).

This also adds unit tests ensuring that the REM is executed.


Former-commit-id: f9c02b4a0ff1662b5920c4973f0b8f0239fc6104