CXF-7888 - Handle EscapeHandler for IBM XLXP JAXB implementation.
This allows the escape handler to be set on an IBM XML Marshaller for
environments like WebSphere Liberty where the default com.sun.xml
implementation is replaced with a different JAXB impl. The new tests
verifies that it works for both the com.sun.xml and com.ibm.xml impls.

Former-commit-id: 0dbe05e16e3cd0240ba781ff2fb9e78c1833fcc6