Get rid of IndexArrayOutOfBounds
If an environment defines the java.version with 1.8.0 this code will die with an ArrayIndexOutOfBoundsEx. 

To evade this either the length of the version should be checked for being >6 or, as proposed in the pr, the 1.8.0 version format should be checked for an additional underscore - if its not there then the environment uses something different than the expected value like 1.8.0_161. But it should never die with an AIOB-Ex. Thank you.

Former-commit-id: dfa3dfea750c4d96e46781c6f2f8b5a183b9c97c